requirements.txt
streamlit>=1.20.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
open_clip_torch>=2.20.0
timm>=0.9.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0
